refactor(hooks): useOverlay 스타일/로직 분리
